Modify the types of some fields in Soroban Request/Response. (#560)
| field                                   | before  | now  |
| --------------------------------------- | ------- | ---- |
| GetEventsRequest.startLedger            | String  | Long |
| GetEventsResponse.EventInfo.ledger      | Integer | Long |
| GetLatestLedgerResponse.protocolVersion | Integer | Long |

+++ src/main/java/org/stellar/sdk/responses/sorobanrpc/GetEventsResponse.java
@@ -23,7 +23,7 @@ public class GetEventsResponse {
   public static class EventInfo {
     EventFilterType type;
 
-    Integer ledger;
+    Long ledger;
 
     String ledgerClosedAt;
